move prover service to its own crate
