+++ bin/citrea/Cargo.toml
@@ -23,6 +23,7 @@ citrea-risc0-bonsai-adapter = { path = "../../crates/risc0-bonsai", features = [
 citrea-sequencer = { path = "../../crates/sequencer" }
 citrea-stf = { path = "../../crates/citrea-stf", features = ["native"] }
 ethereum-rpc = { path = "../../crates/ethereum-rpc" }
+prover-services = { path = "../../crates/prover-services" }
 sequencer-client = { path = "../../crates/sequencer-client" }
 
 # Sovereign-SDK deps
